+++ lib/aspera/cli/plugins/orchestrator.rb
@@ -56,8 +56,8 @@ module Aspera
           }
         end
 
-        option :result,      description: "Specify result value as: 'work_step:parameter'", deprecation: 'use keys `step` and `variable` of argument `execution` of `workflows start`'
-        option :synchronous, description: 'Wait for completion', allowed: Type::BOOLEAN, deprecation: 'use key `synchronous` of argument `execution` of `workflows start`'
+        option :result,      description: "Specify result value as: 'work_step:parameter'", deprecation: {last: '4.27.2', message: 'use keys `step` and `variable` of argument `execution` of `workflows start`'}
+        option :synchronous, description: 'Wait for completion', allowed: Type::BOOLEAN, deprecation: {last: '4.27.2', message: 'use key `synchronous` of argument `execution` of `workflows start`'}
         option :ret_style,   description: 'How return type is requested in api', allowed: %i[header arg ext], default: :arg
         option :auth_style,  description: 'Authentication type', allowed: %i[arg_pass head_basic apikey], default: :head_basic